requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pgvector
python-dotenv
pyjwt
bcrypt
groq
python-jose[cryptography]
email-validator
pydantic==2.8.0 
huggingface-hub==0.17.3
sentence-transformers==2.2.2